backend/requirements.txt
django==4.2
djangorestframework
django-cors-headers
drf-yasg
psycopg2-binary
django-environ
